requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3

# Groq
groq>=0.4.2

# Utilities
python-dotenv==1.0.1
sqlalchemy>=2.0.0
huggingface_hub>=0.21.4
requests>=2.31.0
# Dev/Test
pytest>=7.0
